#21 feat : useHideScroll 프로토타입 구현 - 주요 기능   - 매개변수     - isScrollNeedsDisplay(boolean, 옵션): useEffect를 사용해 스크롤바를 자동으로 숨기고 싶은 경우 전달   - Returns     - isScrollCaptured(boolean): ::webkit-scrollbar-thumb 요소의 visibility 등의 속성 값의 flag로 사용 - onScrollCaptured 이벤트에 true     - onScrollCapture: onScrollCapture 이벤트의 핸들러로 사용 - scrollcapture 이벤트가 발생했을 때 isScrollCaptured의 값을 true로 설정     - onScroll: onScroll 이벤트의 핸들러로 사용 - scroll이벤트가 발생했을 때 isScrollCaptured의 값을 false로 바꿔 스크롤바를 숨김
